Module dos_actors::clients

Actors clients

The module holds the implementation of the different clients that can be assigned to Actors.

Any structure can become a client to an Actor if it implements the Update trait with either or both Read and Write traits.

Example

Logging

A simple logger with a single entry:

use dos_actors::prelude::*;
let logging = Logging::<f64>::default();

A logger with 2 entries and pre-allocated with 1000 elements:

use dos_actors::prelude::*;
let logging = Logging::<f64>::default().n_entry(2).capacity(1_000);

Signals

A constant signal for 100 steps

use dos_actors::prelude::*;
let signal = Signals::new(1, 100).signals(Signal::Constant(3.14));

A 2 outputs signal made of a constant and a sinusoid for 100 steps

use dos_actors::prelude::*;
let signal = Signals::new(2, 100)
               .output_signal(0, Signal::Constant(3.14))
               .output_signal(1, Signal::Sinusoid{
                                        amplitude: 1f64,
                                        sampling_frequency_hz: 1000f64,
                                        frequency_hz: 20f64,
                                        phase_s: 0f64
               });

Rate transitionner

A sample-and-hold rate transition for a named output/input pair sampling a Vec

use dos_actors::prelude::*;
enum MyIO {};
let sampler = Sampler::<Vec<f64>, MyIO>::default();

Modules

arrow_client

Actor client for Apache Arrow

ceo

fem

GMT Finite Element Model client

fsm

GMT Fast Steering Mirror control clients

lom

Linear Optical Model client

m1

GMT M1 control clients

mount

GMT mount control model

windloads

CFD wind loads client implementation

Structs

Logging

Simple data logging

Sampler

Sample-and-hold rate transitionner

Signals

Signals generator

Enums

Signal

Signal types