Module dos_actors::model

Integrated model

The module implements the high-level integrated model interface. The model is build from a collection of actors.

The model has 4 states:

Unknown: model state at its creation

Ready: model state after succesfully performing runtime checks on inputs and outputs on all the actors, the model can move to the Ready state only from the Unknown state

Running: model state while all the actors are performing their respective tasks, the model can move to the Running state only from the Ready state

Completed: model state after the succesful completion of the tasks of all the actors, the model can move to the Completed state only from the Running state

Example

A 3 actors model with Signals, Sampler and Logging clients is build with:

use dos_actors::prelude::*;
let mut source: Initiator<_> = Signals::new(1, 100).into();
enum Source {};
let mut sampler: Actor<_, 1, 10> = Sampler::<Vec<f64>, Source>::default().into();
let logging = Logging::<f64>::default().into_arcx();
let mut sink = Terminator::<_, 10>::new(logging);

sampler decimates source with a 1:10 ratio. The source connects to the sampler using the empty enum type Source as the data identifier. The source data is then logged into the client of the sink actor.

source.add_output().build::<Vec<f64>, Source>().into_input(&mut sampler);
sampler.add_output().build::<Vec<f64>,Source>().into_input(&mut sink);

A model is build from the set of actors:

Model::new(vec![Box::new(source), Box::new(sampler), Box::new(sink)]);

Actors are checked for inputs/outputs consistencies:

Model::new(vec![Box::new(source), Box::new(sampler), Box::new(sink)])
       .check()?;

The model run the actor tasks:

Model::new(vec![Box::new(source), Box::new(sampler), Box::new(sink)])
       .check()?
       .run();

and wait for the tasks to finish:

Model::new(vec![Box::new(source), Box::new(sampler), Box::new(sink)])
       .check()?
       .run()
       .wait()
       .await?;

Once the model run to completion, the data from logging is read with:

let data  = &*(*logging.lock().await);

Structs

Model

Actor model

Enums

Completed

ModelError

Ready

Running

Unknown