Trait dos_actors::Task

pub trait Task: Send {
    fn async_run<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn spawn(self) -> JoinHandle<()>;

    fn check_inputs(&self) -> Result<()>;

    fn check_outputs(&self) -> Result<()>;

    fn task<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
}

Required methods

fn async_run<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = Result<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Runs the Actor infinite loop

The loop ends when the client data is None or when either the sending of receiving end of a channel is dropped

fn spawn(self) -> JoinHandle<()>

Run the actor loop in a dedicated thread

fn check_inputs(&self) -> Result<()>

Validates the inputs

Returns en error if there are some inputs but the inputs rate is zero or if there are no inputs and the inputs rate is positive

fn check_outputs(&self) -> Result<()>

Validates the outputs

Returns en error if there are some outputs but the outputs rate is zero or if there are no outputs and the outputs rate is positive

fn task<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = ()> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Run the actor loop